Stop managing connections count when max=infinity
Whenever the max connections is set to infinity, the connections
counting key is not created, or is deleted. When using a numeric value,
the connection count is created or maintained.

Moreover, trying to reduce a listener's counter while the max connection
number is set to `infinity` will return 0 and avoid all
operations as they are meaningless.